Move test helpers to tests/e2e/utils/ package
Move expected_lanes.py and fbc_parser.py into tests/e2e/utils/ to
separate test utilities from test files.
