+++ tests/e2e/utils/fbc_parser.py
